AI Web Research Assistant/requirements.txt
langchain
langgraph
langchain-community
langchain-openai
langchain-google-genai
langchain-groq
python-dotenv
langchain-ollama
ipykernel
huggingface_hub
sentence-transformers
chromadb
langchain-text-splitters
streamlit
python-dotenv
streamlit
unstructured
tiktoken
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
protobuf
chromadb
sentence-transformers